More baseline stuff
Starting to get a better idea of what's working and what isn't.

Will likely be moving to a mutable game state. Will also probably redesign
actions. I'm going to focus more on the night phase first because I think
it's going to drive most of the complexity of the game design, which the
day phase will depend on.